# backend/docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ALPACART API",
    "description": "API Backend de Alpacart — Marca premium de alpaca peruana",
    "version": "1.0.0",
    "contact": {
      "name": "Alpacart Textiles",
      "url": "https://alpacart.com"
    }
  },
  "servers": [
    { "url": "http://localhost:8000/api/v1", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearer": [] }],
  "paths": {
...
    "/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "Listar pedidos",
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "perPage", "in": "query", "schema": { "type": "integer", "default": 25 } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Lista paginada de pedidos" } }
      },
      "post": {
        "tags": ["Orders"],
        "summary": "Crear pedido (admin)",
        "security": [{ "bearer": [] }],
        "responses": { "201": { "description": "Pedido creado" } }
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Detalle del pedido",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Detalle del pedido" } }
      }
    },
    "/orders/{id}/status": {
      "put": {
        "tags": ["Orders"],
        "summary": "Cambiar estado",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Estado actualizado" } }
      }
    },
    "/orders/{id}/notes": {
      "post": {
        "tags": ["Orders"],
        "summary": "Agregar nota",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "201": { "description": "Nota agregada" } }
      }
    },
    "/orders/{id}/events": {
      "get": {
        "tags": ["Orders"],
        "summary": "Timeline del pedido",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Eventos del pedido" } }
      }
    },
...
    "/create-payment-intent": {
      "post": {
        "tags": ["Payments"],
        "summary": "Crear Payment Intent (Stripe)",
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "orderId": { "type": "string" },
                  "amount": { "type": "number" },
                  "currency": { "type": "string", "default": "usd" },
                  "customerEmail": { "type": "string", "format": "email" }
                },
                "required": ["orderId", "amount", "currency"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Payment Intent creado" } }
      }
    },
    "/transactions": {
      "get": {
        "tags": ["Payments"],
        "summary": "Listar transacciones",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Lista de transacciones" } }
      }
    },
    "/transactions/{id}/refund": {
      "post": {
        "tags": ["Payments"],
        "summary": "Reembolsar transacción",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "201": { "description": "Reembolso procesado" } }
      }
    },
...
    "/reservations/release-expired": {
      "post": {
        "tags": ["Payments"],
        "summary": "Liberar reservas expiradas",
        "security": [{ "bearer": [] }],
        "responses": { "201": { "description": "Reservas liberadas" } }
      }
    },
...
  }
}